@@ -1,6 +1,6 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { NavMenuComponent } from './nav-menu/nav-menu.component';
@@ -10,25 +10,17 @@ import { NewCertificateComponent } from './new-certificate/new-certificate.compo
 import { RouterModule } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    NavMenuComponent,
-    HomeComponent,
-    CertificatesComponent,
-    NewCertificateComponent
-  ],
-  imports: [
-    //BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
-    HttpClientModule,
-    FormsModule,
-    ReactiveFormsModule,
-    RouterModule.forRoot([
-      { path: '', component: CertificatesComponent, pathMatch: 'full' },
-      { path: 'new-certificate', component: NewCertificateComponent },
-    ])
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        NavMenuComponent,
+        HomeComponent,
+        CertificatesComponent,
+        NewCertificateComponent
+    ],
+    bootstrap: [AppComponent], imports: [FormsModule,
+        ReactiveFormsModule,
+        RouterModule.forRoot([
+            { path: '', component: CertificatesComponent, pathMatch: 'full' },
+            { path: 'new-certificate', component: NewCertificateComponent },
+        ])], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
